# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LQGovernance — Board Management Portal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /boards:
    get:
      operationId: listBoards
      tags: [boards]
      summary: List boards accessible to current user
      responses:
        "200":
          description: List of boards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Board"
    post:
      operationId: createBoard
      tags: [boards]
      summary: Create a new board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBoardBody"
      responses:
        "201":
          description: Created board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
  /boards/{id}:
    get:
      operationId: getBoard
      tags: [boards]
      summary: Get a board by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Board with members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardWithMembers"
    patch:
      operationId: updateBoard
      tags: [boards]
      summary: Update board settings (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBoardBody"
      responses:
        "200":
          description: Updated board
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Board"
  /boards/{id}/members:
    get:
      operationId: getBoardMembers
      tags: [boards]
      summary: Get members of a board
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Board members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BoardMember"
    post:
      operationId: addBoardMember
      tags: [boards]
      summary: Add a member to a board
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBoardMemberBody"
      responses:
        "201":
          description: Member added
  /boards/{id}/members/{personId}:
    patch:
      operationId: updateBoardMember
      tags: [boards]
      summary: Update a board member's role and/or voting weight (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: personId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBoardMemberBody"
      responses:
        "200":
          description: Member updated
    delete:
      operationId: removeBoardMember
      tags: [boards]
      summary: Remove a member from a board (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: personId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Member removed

  # Meetings
...
  /audit:
    get:
      operationId: listAuditEntries
      tags: [audit]
      summary: List audit-trail entries (admin only)
      parameters:
        - name: personId
          in: query
          required: false
          schema:
            type: string
        - name: action
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Audit entries, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEntry"
...
  /audit/verify:
    get:
      operationId: verifyAuditChain
      tags: [audit]
      summary: Verify the audit hash chain (admin only)
      description: >
        Replays the audit hash chain and reports the first broken link. Detects a
        naive edit; a full re-seal by an actor with database write access is not
        detectable without the external anchor (see SECURITY.md).
      responses:
        "200":
          description: Chain intact
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  count: { type: integer }
        "409":
          description: Chain broken — a link does not verify
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  count: { type: integer }
                  brokenAtIndex: { type: integer }
                  brokenRowId: { type: string }

  # ── Deleted records: recycle bin + restore (issue #11) ──
...
components:
  schemas:
...
    Person:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, member, observer, management]
        title:
          type: ["string", "null"]
        avatarColor:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - createdAt
...
    Board:
      type: object
      properties:
        id:
          type: string
        organizationId:
          type: ["string", "null"]
        name:
          type: string
        abbreviation:
          type: ["string", "null"]
        type:
          type: string
          enum: [board, committee]
        proxyLimit:
          type: integer
          description: Max proxies one member may hold on a single vote (0 disables proxy voting).
        memberCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - proxyLimit
        - memberCount
        - createdAt
    BoardWithMembers:
      type: object
      properties:
        id:
          type: string
        organizationId:
          type: ["string", "null"]
        name:
          type: string
        abbreviation:
          type: ["string", "null"]
        type:
          type: string
        proxyLimit:
          type: integer
        members:
          type: array
          items:
            $ref: "#/components/schemas/BoardMember"
        createdAt:
          type: string
      required:
        - id
        - name
        - type
        - proxyLimit
        - members
        - createdAt
    BoardMember:
      type: object
      properties:
        id:
          type: string
        personId:
          type: string
        boardId:
          type: string
        roleInBoard:
          type: ["string", "null"]
        votingWeight:
          type: integer
          description: Voting weight of this member on this board (positive integer, default 1).
        person:
          $ref: "#/components/schemas/Person"
      required:
        - id
        - personId
        - boardId
        - votingWeight
        - person
    CreateBoardBody:
      type: object
      properties:
        name:
          type: string
        abbreviation:
          type: ["string", "null"]
        type:
          type: string
          enum: [board, committee]
      required:
        - name
        - type
    UpdateBoardBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        proxyLimit:
          type: ["integer", "null"]
          description: Max proxies one member may hold on a single vote (0 disables proxy voting).
    AddBoardMemberBody:
      type: object
      properties:
        personId:
          type: string
        roleInBoard:
          type: ["string", "null"]
        votingWeight:
          type: ["integer", "null"]
          description: Voting weight (positive integer). Defaults to 1.
      required:
        - personId
    UpdateBoardMemberBody:
      type: object
      properties:
        roleInBoard:
          type: ["string", "null"]
        votingWeight:
          type: ["integer", "null"]
          description: New voting weight (positive integer). Already-cast ballots keep their snapshotted weight.

    # Meeting
...
    AuditEntry:
      type: object
      properties:
        id:
          type: string
        personId:
          type: ["string", "null"]
        action:
          type: string
        actionLabel:
          type: string
        entityType:
          type: ["string", "null"]
        entityId:
          type: ["string", "null"]
        details:
          type: ["object", "null"]
        ipAddress:
          type: ["string", "null"]
        createdAt:
          type: string
        person:
          oneOf:
            - $ref: "#/components/schemas/AuditPerson"
            - type: "null"
      required:
        - id
        - action
        - actionLabel
        - createdAt
    AuditPerson:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatarColor:
          type: ["string", "null"]
      required:
        - id
        - name
        - email
        - role

    # ── Deleted records (recycle bin, issue #11) ──
...